Move types/app.ts to file-system/models
