+++ src/components/ai-chat/types.ts
@@ -1,5 +1,5 @@
 import type React from "react";
-import type { FileEntry } from "../../types/app";
+import type { FileEntry } from "../../file-system/models/app";
 import type { Buffer } from "../../types/buffer";
 
 interface ToolCall {
