+++ src/components/editor/breadcrumb.tsx
@@ -5,9 +5,9 @@ import { EDITOR_CONSTANTS } from "../../constants/editor-constants";
 import FileIcon from "../../file-explorer/views/file.icon";
 import { readDirectory } from "../../file-system/controllers/platform";
 import { useFileSystemStore } from "../../file-system/controllers/store";
+import type { FileEntry } from "../../file-system/models/app";
 import { useBufferStore } from "../../stores/buffer-store";
 import { useUIState } from "../../stores/ui-state-store";
-import type { FileEntry } from "../../types/app";
 
 export default function Breadcrumb() {
   const buffers = useBufferStore.use.buffers();
