+++ src/components/layout/main-sidebar.tsx
@@ -4,12 +4,12 @@ import { useMemo } from "react";
 import { cn } from "@/utils/cn";
 import FileTree from "../../file-explorer/views/file-tree";
 import { useFileSystemStore } from "../../file-system/controllers/store";
+import type { FileEntry } from "../../file-system/models/app";
 import { usePersistentSettingsStore } from "../../settings/stores/persistent-settings-store";
 import { useBufferStore } from "../../stores/buffer-store";
 import { useProjectStore } from "../../stores/project-store";
 import { useSidebarStore } from "../../stores/sidebar-store";
 import { useUIState } from "../../stores/ui-state-store";
-import type { FileEntry } from "../../types/app";
 import GitView from "../git/git-view";
 import RemoteConnectionView from "../remote/remote-connection-view";
 import SearchView from "../search-view";
