+++ src/file-system/controllers/utils.ts
@@ -1,5 +1,5 @@
-import type { FileEntry } from "../../types/app";
 import { sortFileEntries } from "../../utils/file-tree-utils";
+import type { FileEntry } from "../models/app";
 
 // Common directories and patterns to ignore for performance
 const IGNORE_PATTERNS = [
