+++ src/stores/ai-chat/types.ts
@@ -1,5 +1,5 @@
 import type { Chat, Message } from "../../components/ai-chat/types";
-import type { FileEntry } from "../../types/app";
+import type { FileEntry } from "../../file-system/models/app";
 
 export interface AIChatState {
   // Input state
